refactor(button): remove the size property altogether
