+++ src/components/Button/Button.stories.tsx
@@ -6,7 +6,7 @@ import { generateControl } from '../../utils/tests/storybook';
 import { Inline, Stack } from '../layout';
 import Button from './Button';
 import { ButtonProps } from './Button.types';
-import { ButtonColors, ButtonSizes, ButtonVariants } from './Button.enums';
+import { ButtonColors, ButtonVariants } from './Button.enums';
 import { SpaceSizes } from '../../theme';
 
 export default {
@@ -26,9 +26,6 @@ export default {
     variant: {
       ...generateControl('select', ButtonVariants),
     },
-    size: {
-      ...generateControl('select', ButtonSizes),
-    },
     color: {
       ...generateControl('select', ButtonColors),
     },
@@ -42,7 +39,6 @@ export const Playground: Story<ButtonProps> = (args) => <Button {...args} />;
 Playground.args = {
   children: 'Playground Button',
   variant: ButtonVariants.solid,
-  size: ButtonSizes.md,
   color: ButtonColors.primary,
 };
 Playground.parameters = {
@@ -166,174 +162,47 @@ export const TextButtons: Story = () => (
   </Stack>
 );
 
-export const SizedButtons: Story = () => (
-  <Stack gap={SpaceSizes.lg}>
-    <Inline align="center" gap={SpaceSizes.md}>
-      <Button color="primary" size="lg" variant="solid">
-        Large Button
-      </Button>
-      <Button color="primary" size="md" variant="solid">
-        Medium Button
-      </Button>
-      <Button color="primary" size="sm" variant="solid">
-        Small Button
-      </Button>
-    </Inline>
-    <Inline align="center" gap={SpaceSizes.md}>
-      <Button color="primary" size="lg" variant="outline">
-        Large Button
-      </Button>
-      <Button color="primary" size="md" variant="outline">
-        Medium Button
-      </Button>
-      <Button color="primary" size="sm" variant="outline">
-        Small Button
-      </Button>
-    </Inline>
-    <Inline align="center" gap={SpaceSizes.md}>
-      <Button color="primary" size="lg" variant="text">
-        Large Button
-      </Button>
-      <Button color="primary" size="md" variant="text">
-        Medium Button
-      </Button>
-      <Button color="primary" size="sm" variant="text">
-        Small Button
-      </Button>
-    </Inline>
-  </Stack>
-);
-
 export const LoadingButtons: Story = () => (
-  <Stack gap={SpaceSizes.lg}>
-    <Inline align="center" gap={SpaceSizes.md}>
-      <Button color="primary" size="lg" variant="solid" isLoading>
-        Large Button
-      </Button>
-      <Button color="primary" size="md" variant="solid" isLoading>
-        Medium Button
-      </Button>
-      <Button color="primary" size="sm" variant="solid" isLoading>
-        Small Button
-      </Button>
-    </Inline>
-    <Inline align="center" gap={SpaceSizes.md}>
-      <Button color="primary" size="lg" variant="outline" isLoading>
-        Large Button
-      </Button>
-      <Button color="primary" size="md" variant="outline" isLoading>
-        Medium Button
-      </Button>
-      <Button color="primary" size="sm" variant="outline" isLoading>
-        Small Button
-      </Button>
-    </Inline>
-    <Inline align="center" gap={SpaceSizes.md}>
-      <Button color="primary" size="lg" variant="text" isLoading>
-        Large Button
-      </Button>
-      <Button color="primary" size="md" variant="text" isLoading>
-        Medium Button
-      </Button>
-      <Button color="primary" size="sm" variant="text" isLoading>
-        Small Button
-      </Button>
-    </Inline>
-  </Stack>
+  <Inline gap={SpaceSizes.md}>
+    <Button color="primary" variant="solid" isLoading>
+      Button
+    </Button>
+
+    <Button color="primary" variant="outline" isLoading>
+      Button
+    </Button>
+
+    <Button color="primary" variant="text" isLoading>
+      Button
+    </Button>
+  </Inline>
 );
 
 export const ButtonsWithIconOverview: Story = () => (
-  <Stack gap={SpaceSizes.lg}>
-    <Inline align="center" gap={SpaceSizes.md}>
-      <Button
-        color="primary"
-        iconName={SSCIconNames.wrench}
-        size="lg"
-        variant="solid"
-      >
-        Large Button
-      </Button>
-      <Button
-        color="primary"
-        iconName={SSCIconNames.wrench}
-        size="md"
-        variant="solid"
-      >
-        Medium Button
-      </Button>
-      <Button
-        color="primary"
-        iconName={SSCIconNames.wrench}
-        size="sm"
-        variant="solid"
-      >
-        Small Button
-      </Button>
-    </Inline>
-    <Inline align="center" gap={SpaceSizes.md}>
-      <Button
-        color="primary"
-        iconName={SSCIconNames.wrench}
-        size="lg"
-        variant="outline"
-      >
-        Large Button
-      </Button>
-      <Button
-        color="primary"
-        iconName={SSCIconNames.wrench}
-        size="md"
-        variant="outline"
-      >
-        Medium Button
-      </Button>
-      <Button
-        color="primary"
-        iconName={SSCIconNames.wrench}
-        size="sm"
-        variant="outline"
-      >
-        Small Button
-      </Button>
-    </Inline>
-    <Inline align="center" gap={SpaceSizes.md}>
-      <Button
-        color="primary"
-        iconName={SSCIconNames.wrench}
-        size="lg"
-        variant="text"
-      >
-        Large Button
-      </Button>
-      <Button
-        color="primary"
-        iconName={SSCIconNames.wrench}
-        size="md"
-        variant="text"
-      >
-        Medium Button
-      </Button>
-      <Button
-        color="primary"
-        iconName={SSCIconNames.wrench}
-        size="sm"
-        variant="text"
-      >
-        Small Button
-      </Button>
-    </Inline>
-  </Stack>
+  <Inline gap={SpaceSizes.md}>
+    <Button color="primary" iconName={SSCIconNames.wrench} variant="solid">
+      Button
+    </Button>
+
+    <Button color="primary" iconName={SSCIconNames.wrench} variant="outline">
+      Button
+    </Button>
+
+    <Button color="primary" iconName={SSCIconNames.wrench} variant="text">
+      Button
+    </Button>
+  </Inline>
 );
 
 export const ExpandedButtons: Story = () => (
   <Stack gap={SpaceSizes.lg}>
-    <Button size="md" variant="solid" isExpanded>
+    <Button variant="solid" isExpanded>
       Full width button
     </Button>
-    <Button size="md" variant="outline" isExpanded>
+    <Button variant="outline" isExpanded>
       Full width button
     </Button>
-    <Button size="md" variant="text" isExpanded>
+    <Button variant="text" isExpanded>
       Full width button
     </Button>
   </Stack>
