+++ src/components/Button/Button.types.ts
@@ -1,28 +1,3 @@
-import React from 'react';
-import { To } from 'history'; // just importing types
+import { BaseButtonProps } from '../_internal/BaseButton/BaseButton.types';
 
-import { SpacingSizeValue } from '../../types/spacing.types';
-import { Types as IconTypes, SSCIcons } from '../Icon/Icon.types';
-import { ButtonColors, ButtonSizes, ButtonVariants } from './Button.enums';
-
-export type Variants = typeof ButtonVariants[keyof typeof ButtonVariants];
-export type Sizes = typeof ButtonSizes[keyof typeof ButtonSizes];
-export type Colors = typeof ButtonColors[keyof typeof ButtonColors];
-
-export interface ButtonProps
-  extends Omit<React.HTMLProps<HTMLButtonElement>, 'size' | 'as'> {
-  variant?: Variants;
-  size?: Sizes;
-  color?: Colors;
-  onClick?: React.MouseEventHandler;
-  isDisabled?: boolean;
-  isLoading?: boolean;
-  isExpanded?: boolean;
-  margin?: SpacingSizeValue;
-  href?: string;
-  to?: To;
-  className?: string;
-  iconName?: SSCIcons | string;
-  loadingText?: string;
-  iconType?: IconTypes | string;
-}
+export type ButtonProps = Omit<BaseButtonProps, 'size'>;
