--- src/components/Button/StyledButton.types.ts
@@ -1,7 +1,0 @@
-import { ButtonProps } from './Button.types';
-
-export interface BaseStyledButtonProps extends ButtonProps {
-  isDisabled: never;
-  disabled?: boolean;
-  $hasOnlyIcon?: boolean;
-}
