--- src/components/Button/StyledIcon.types.ts
@@ -1,6 +1,0 @@
-import { BaseIconProps } from '../Icon/Icon.types';
-import { Sizes } from './Button.types';
-
-export interface StyledIconProps extends Required<BaseIconProps> {
-  size: Sizes;
-}
