+++ src/components/ButtonGroup/ButtonGroup.types.ts
@@ -1,9 +1,8 @@
 import { ButtonProps } from '../Button/Button.types';
 
-export type ButtonGroupProps = Pick<ButtonProps, 'variant' | 'size' | 'color'>;
+export type ButtonGroupProps = Pick<ButtonProps, 'variant' | 'color'>;
 
 export type ButtonGroupWrapperProps = {
   $variant: ButtonGroupProps['variant'];
-  $size: ButtonGroupProps['size'];
   $color: ButtonGroupProps['color'];
 };
