+++ src/components/Drawer/Drawer.stories.tsx
@@ -41,10 +41,8 @@ export default {
 function Footer() {
   return (
     <Inline gap={SpaceSizes.md} justify="flex-end">
-      <Button size={SpaceSizes.lg} variant="outline">
-        Cancel
-      </Button>
-      <Button size={SpaceSizes.lg}>Confirm</Button>
+      <Button variant="outline">Cancel</Button>
+      <Button>Confirm</Button>
     </Inline>
   );
 }
@@ -74,12 +72,7 @@ export const Playground: Story<DrawerProps> = (args) => {
   return (
     <Padbox paddingSize={SpaceSizes.xxl}>
       <Inline justify="center">
-        <Button
-          size={ButtonEnums.ButtonSizes.lg}
-          onClick={() => setIsShown(!isShown)}
-        >
-          Open drawer
-        </Button>
+        <Button onClick={() => setIsShown(!isShown)}>Open drawer</Button>
         {isShown && (
           <Drawer
             footer={<Footer />}
