+++ src/components/SemanticModal/SemanticModal.tsx
@@ -63,7 +63,6 @@ const SemanticModal = forwardRef<HTMLDivElement, SemanticModalProps>(
         color={color}
         href={(action as AbsoluteLinkActionKind<[React.MouseEvent]>).href}
         name={action.name}
-        size={SpaceSizes.lg}
         to={(action as RelativeLinkActionKind<[React.MouseEvent]>).to}
         variant={buttonVariant}
         onClick={action.onClick}
