+++ src/components/_internal/BaseButton/BaseButton.enums.ts
@@ -1,0 +1,17 @@
+export const BaseButtonVariants = {
+  solid: 'solid',
+  outline: 'outline',
+  text: 'text',
+} as const;
+
+export const BaseButtonColors = {
+  primary: 'primary',
+  secondary: 'secondary',
+  danger: 'danger',
+} as const;
+
+export const BaseButtonSizes = {
+  lg: 'lg',
+  md: 'md',
+  sm: 'sm',
+} as const;
