+++ src/components/_internal/BaseButton/BaseStyledButton.types.ts
@@ -1,0 +1,7 @@
+import { BaseButtonProps } from './BaseButton.types';
+
+export interface BaseStyledButtonProps extends BaseButtonProps {
+  isDisabled: never;
+  disabled?: boolean;
+  $hasOnlyIcon?: boolean;
+}
