+++ src/components/_internal/BaseButton/BaseStyledIcon.types.ts
@@ -1,0 +1,6 @@
+import { BaseIconProps } from '../../Icon/Icon.types';
+import { Sizes } from './BaseButton.types';
+
+export interface BaseStyledIconProps extends Required<BaseIconProps> {
+  size: Sizes;
+}
