+++ src/components/_internal/BaseButton/index.ts
@@ -1,0 +1,2 @@
+export * as BaseButtonEnums from './BaseButton.enums';
+export { default as BaseButton } from './BaseButton';
